+++ src/ur5e_controller/ur5e_thread.cpp
@@ -23,7 +23,7 @@ void ur5e_controller() {
   // 完成初始化
   time = get_current_time();
   pid_t urPid = get_tid();
-  ROS_INFO("T[%Lf] P[%d] UR5e thread is Ready!", time, urPid);
+  ROS_INFO("P[%d] T[%Lf] UR5e thread is Ready!", time, urPid);
 
   /* 等待线程同步 */
   ts = threadmanager.wait_for_syc();